ETS-38 - Call GIAS SOAP APi

+++ src/Dfe.Spi.GiasAdapter.Domain/GiasApi/Establishment.cs
@@ -2,7 +2,7 @@ namespace Dfe.Spi.GiasAdapter.Domain.GiasApi
 {
     public class Establishment
     {
-        public int Urn { get; set; }
+        public long Urn { get; set; }
         public string Name { get; set; }
     }
 }
